Добавление нового товара

+++ src/main/java/com/plantshop/repository/ProductRepository.java
@@ -3,5 +3,6 @@ package com.plantshop.repository;
 import com.plantshop.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
+    //public Product delete(Integer id);
 }
